+++ tests/CacheOrchestrator.AdminConsole.Tests/HintRuleRegistryAndDisableStoreTests.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 
-namespace CacheOrchestrator.UnitTests.Admin;
+namespace CacheOrchestrator.AdminConsole.Tests;
 
 public class HintRuleRegistryAndDisableStoreTests
 {